refactor(client): extract settings locale dictionaries
Keep the settings view as presentation-only and share locale keys from a
dedicated module.
